Move `expect` duplicate method into a new composable type (FeedLoaderTestCase) that can be used across XCTestCase classes
